feat(security): centralize Trusted Types handling and fix icon innerHTML usages
Introduced a shared html utility for Trusted Types and replaced verbose innerHTML assignments in UI components with setTrustedInnerHtml.
